nlp/translation/requirements.txt
# Neural Network Model Dependencies
pytorch
torchmetrics

# Core Python Libraries
numpy

# Visualization Tools
tensorboard

# huggingface datasets
datasets
tokenizers
transformers
sacrebleu
evaluate
sentencepiece
sacremoses
accelerate>=0.26.0